requirements.txt
alembic==1.13.1
annotated-types==0.6.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
colorama==0.4.6
greenlet==3.0.3
loguru==0.7.2
Mako==1.3.3
MarkupSafe==2.1.5
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
TgCrypto==1.2.5
typing_extensions==4.11.0
win32-setctime==1.1.0
